feat: add flexible Material → Variant → Option pricing model
- New entities: Material, MaterialVariant, MaterialOption with cascade relationships
- QuoteItem now supports both products and material options (nullable FKs)
- Quantity changed from int to decimal to support fractional amounts (e.g., m²)
- QuoteItem stores Description snapshot for historical readability
- New MaterialRepository and MaterialService following existing patterns
- MaterialListView: grid showing all materials with variant/option counts
- MaterialFormView: TreeView-based editor for variants and options
- QuoteFormView: dual-mode item selection (product/service OR material)
- Updated DashboardView with Materiales card button
- Updated QuoteDetailView and PdfExportService for new Description field
- Seed data includes 8 materials with real pricing examples
- DB is deleted and re-seeded on launch (development schema reset)

+++ Services/QuoteCalculationService.cs
@@ -4,7 +4,7 @@ namespace SistemaCotizaciones.Services
 {
     public class QuoteCalculationService
     {
-        public decimal CalculateSubtotal(int quantity, decimal unitPrice)
+        public decimal CalculateSubtotal(decimal quantity, decimal unitPrice)
         {
             return quantity * unitPrice;
         }
